# Conflicts:
#	resources/views/pages/product-details.blade.php
#	resources/views/welcome.blade.php

+++ routes/web.php
@@ -4,10 +4,17 @@ use App\Mail\VerificationMail;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Route;
+Route::get('link', function (){
+    Artisan::call('storage:link');
+});
 
+Route::get('clear', function (){
+    Artisan::call('cache:clear');
+});
 Route::get('/createslug', function (){
     $categories = \App\Models\Brand::all();
 
@@ -94,7 +101,7 @@ Route::get('/faq',function(){
 })->name('faq');
 
 Route::get('/',  'WelcomeController@index')->name('home');
-Route::post('layouts/', 'Users\NavbarController@store')->name('pages.search');
+Route::get('layouts/', 'Users\NavbarController@store')->name('pages.search');
 Route::get('checkout', 'Users\CheckoutController@index')->middleware(['auth'])->name('checkout');
 Route::post('checkout', 'Users\CheckoutController@checkout')->name('checkout.final');
 Route::get('checkoutconfirm', 'Users\CheckoutController@checkoutConfirm')->name('checkout.confirm');
@@ -109,12 +116,12 @@ Route::get('brands', 'Users\ContactController@brandShow')->name('brands.show');
 Route::get('brands/{slug}', 'Users\NavbarController@brandProduct')->name('brandProduct.show');
 
 //user profile
-Route::get('profile', 'Users\NavbarController@profile')->name('profile.show');
-Route::get('profile/{user}/edit', 'Users\NavbarController@profileEdit')->name('profile.edit');
-Route::put('profile/{user}', 'Users\NavbarController@profileUpdate')->name('profile.update');
+Route::get('profile', 'Users\NavbarController@profile')->middleware('auth:web')->name('profile.show');
+Route::get('profile/{user}/edit', 'Users\NavbarController@profileEdit')->middleware('auth:web')->name('profile.edit');
+Route::put('profile/{user}', 'Users\NavbarController@profileUpdate')->middleware('auth:web')->name('profile.update');
 
 // Cancel Order
-Route::delete('profile/{order}', 'Users\NavbarController@orderCancel')->name('profile.order.cancel');
+Route::delete('profile/{order}', 'Users\NavbarController@orderCancel')->middleware(['auth'])->name('profile.order.cancel');
 
 //vendor page show
 Route::get('allvendor', 'Users\VendorProductsController@allVendor')->name('allVendor.show');
@@ -130,6 +137,10 @@ Route::get('blogall', 'Users\BlogController@allBog')->name('blog.allBog');
 Route::get('blogall/{slug}', 'Users\BlogController@show')->middleware(['auth'])->name('blog.show');
 Route::DELETE('blogDelete/{id}', 'Users\BlogController@destroy')->middleware(['auth'])->name('blog.destroy');
 
+// promotion route
+Route::get('promotion', 'WelcomeController@promotion')->name('promotion.category');
+Route::get('promotion/{event_id?}/{category_id}', 'WelcomeController@promotionProduct')->name('promotion.products');
+Route::get('changelocation/{region}/{id}', 'WelcomeController@changeLocation');
 
 
 // comment route
@@ -151,6 +162,14 @@ Route::prefix('pages')->group(function(){
 
 //User Auth
 Auth::routes();
+Route::prefix('admin')->group(function (){
+    Route::get('/', 'Users\Admin\AdminController@index')->name('admin.dashboard');
+    Route::get('/login', 'Auth\AdminLoginController@showLoginForm')->name('admin.login');
+    Route::post('/login', 'Auth\AdminLoginController@login')->name('admin.login.submit');
+    Route::get('/register', 'Auth\AdminRegisterController@showRegisterForm')->name('admin.register');
+    Route::post('/register', 'Auth\AdminRegisterController@register')->name('admin.register.submit');
+    Route::post('/logout', 'Auth\AdminLoginController@logout')->name('admin.logout');
+});
 Route::get('/email/verify', function () {
     return view('auth.verify');
 })->name('verification.verify');
@@ -166,13 +185,7 @@ Route::get('verify', 'Auth\RegisterController@verify')->name('verify.mail');
 
 
 // Admin Auth routes
-Route::prefix('admin')->group(function(){
-    Route::get('/', 'Users\Admin\AdminController@index')->name('admin.dashboard');
-    Route::get('/login', 'Auth\AdminLoginController@showLoginForm')->name('admin.login');
-    Route::post('/login', 'Auth\AdminLoginController@login')->name('admin.login.submit');
-    Route::get('/register', 'Auth\AdminRegisterController@showRegisterForm')->name('admin.register');
-    Route::post('/register', 'Auth\AdminRegisterController@register')->name('admin.register.submit');
-    Route::post('/logout', 'Auth\AdminLoginController@logout')->name('admin.logout');
+Route::prefix('admin')->middleware('auth:admin')->group(function(){
     Route::get('allOrders/{id}', 'Users\Admin\OrderController@allOrders')->name('orders.allOrders');
     Route::get('allevent/{id}', 'Users\Admin\EventProductController@allEventProducts')->name('event.allEvents');
     Route::get('sellreport', 'Users\Admin\AdminController@adminMonthlySell')->name('sell.report');
@@ -196,7 +209,7 @@ Route::prefix('admin')->group(function(){
         Route::post('vendorChange', 'AdminController@change')->name('vendor.change.status');
 
         //category by product, subcategory
-        Route::post('productbycat/{category}', 'CategoryController@productByCategory')->name('product.by.cat');
+        Route::post('productbycat/{secondarySubCategory}', 'SecondarySubCategoryController@productBySecondSubCategory')->name('product.by.cat');
         Route::post('subcatbycat/{category}', 'CategoryController@subCategoryByCategory')->name('sub.cat.by.cat');
         Route::post('secondsubcatbysubcat/{subcategory}', 'SubCategoryController@secondarySubBySubCategory')->name('second.sub.cat.by.sub.cat');
 
@@ -214,6 +227,7 @@ Route::prefix('admin')->group(function(){
             Route::get('user', 'UserController@allUser')->name('user.no.order');
 
             Route::get('contactusslider', 'ContactUsSliderController@allSlider')->name('slider');
+            Route::get('categorySlider', 'CategorySliderController@allCategorySlider')->name('categorySlider');
             Route::get('area', 'AreaController@allArea')->name('allArea');
             Route::get('emi', 'EMIController@withoutAdmin')->name('emi');
             Route::get('event', 'EventController@allEvent')->name('event');
@@ -228,7 +242,7 @@ Route::prefix('admin')->group(function(){
 
 });
 
-Route::prefix('admin')->namespace('Users\Vendor')->group(function (){
+Route::prefix('admin')->namespace('Users\Vendor')->middleware('auth:admin')->group(function (){
 
     Route::prefix('vendor')->group(function(){
         Route::resource('productCapacity', 'ProductCapacityController');
@@ -243,7 +257,7 @@ Route::prefix('admin')->namespace('Users\Vendor')->group(function (){
     });
 });
 
-Route::prefix('admin')->group(function (){
+Route::prefix('admin')->middleware('auth:admin')->group(function (){
     Route::get('users', 'Users\Admin\UserController@index')->name('admin.all.users');
     Route::get('change/{user}/{currentStatus}', 'Users\Admin\UserController@changeStatus')->name('admin.all.users.change.status');
 
@@ -252,7 +266,7 @@ Route::prefix('admin')->group(function (){
     Route::delete('blog/{blog}', 'Users\BlogController@destroy')->name('blog.destroy');
 });
 
-Route::prefix('admin')->namespace('Users\Admin')->group(function(){
+Route::prefix('admin')->middleware('auth:admin')->namespace('Users\Admin')->group(function(){
     Route::resource('category', 'CategoryController');
     Route::resource('subcategory', 'SubCategoryController');
     Route::resource('brand', 'BrandController');
@@ -274,6 +288,7 @@ Route::prefix('admin')->namespace('Users\Admin')->group(function(){
     Route::resource('division', 'DivisionController');
     Route::resource('district', 'DistrictController');
     Route::resource('city', 'CityController');
+    Route::resource('categorySlider', 'CategorySliderController');
 
 
     Route::get('expresswish', 'AdminController@expressWish')->name('admin.express.wish');
